changed big enum to interface
